pyproject.toml
[tool.poetry]
name = "naane"
version = "0.1.0"
description = "NAAN editor with no database"
authors = ["datadavev <605409+datadavev@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.58.1"
fastapi = "^0.95.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
